% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{drop_flagged}
\alias{drop_flagged}
\title{Remove flagged records}
\usage{
drop_flagged(dat)
}
\arguments{
\item{dat}{Tibble from \code{\link[=sense_check]{sense_check()}}.}
}
\value{
\code{dat} with flagged rows removed.
}
\description{
Removes all rows where \code{sense_flag} is not \code{NA} (i.e., rows flagged by
\code{\link[=sense_check]{sense_check()}}).
}
