% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomachr-package.R
\docType{package}
\name{stomachr-package}
\alias{stomachr}
\alias{stomachr-package}
\title{stomachr: Download and Clean ICES Stomach Content Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tidy pipeline for downloading and cleaning ICES stomach content data from the four-table relational schema used by the ICES stomach content database (stomachdata.ices.dk). Produces a single flat tibble suitable for individual-level diet analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxlindmark/stomachr}
  \item \url{https://maxlindmark.github.io/stomachr/}
  \item Report bugs at \url{https://github.com/maxlindmark/stomachr/issues}
}

}
\author{
\strong{Maintainer}: Max Lindmark \email{max.lindmark@slu.se}

}
\keyword{internal}
